src/search/requirements.txt
faiss-cpu>=1.7.2
numpy>=1.24.0